PointHandler: don't deactivate because of synth-mouse events
This is necessary after 0b7ee7a48f5f32fc916cc1914b110bdf68e37486.

If QQuickPointHandler::wantsEventPoint() ever returns false,
QQuickPointerHandler::handlePointerEvent() calls setActive(false).
If AA_SynthesizeMouseForUnhandledTabletEvents is enabled, and
PointHandler is the only handler trying to handle tablet events, it does
not accept the events; so QGuiApplicationPrivate::processTabletEvent()
sends a synth-mouse press for each tablet press. If the position doesn't
match the last-known mouse position, QGuiAppPriv::processMouseEvent()
sends a synth-mouse move, too. PointHandler doesn't care about either of
these synth-mouse events, but it has to say that it "wants" them to
avoid getting deactivated. But in spite of that, we don't want
QQuickSinglePointHandler::handlePointerEventImpl() to call
QQuickHandlerPoint::reset() with those synth-mouse events, because
then the tablet-specific data from the previous QTabletEvent would
get overwritten. In general, pointer handlers handle original
device-specific events only, and we just have to avoid letting these
synth-mouse events interfere.

There were a couple of places that were assuming that if an event
is not a QTouchEvent, then it's safe to cast it to QSinglePointEvent;
that's usually true, but isSinglePointEvent() exists since qtbase
9a1a15b42fb526ad4f80944afb7761bfff1b5c9d so we might as well use it.

Previously, we'd set acceptedButtons: Qt.LeftButton | Qt.RightButton
to indicate that the PointHandler should respond to either one; but
now it's better to set Qt.NoButton to indicate that it doesn't care.
The synth-mouse move has no buttons set; if PointHandler only "wants"
events with specific buttons, that buttonless move will deactivate it
temporarily, and then the synth-mouse move would re-activate it, causing
extra activeChanged and pointChanged signals. But so far, its default
for acceptedButtons is still Qt.LeftButton. Perhaps we should change it.

The special meaning in QQPointerDeviceHandler::wantsPointerEvent() for
Qt.NoButton started already with HoverHandler in
1982d1b1aa55ae44a1a775a5745e5c2f11001398 and then
fc636af3a723ee8b4ee42cf71864ae0df5ca4621 for PinchHandler. Those
handlers fundamentally don't need to care about buttons, so they declare
it by setting acceptedButtons = Qt::NoButton in their ctors, and the
user should not set it differently; whereas with PointHandler, we do
expect the user to set acceptedButtons.

tst_PointHandler::tabletStylus now has code to test both with and
without AA_SynthesizeMouseForUnhandledTabletEvents. But in the "true"
case (the default), it won't pass until we also change
QGuiApplicationPrivate::processMouseEvent() to send synth-mouse-moves
from the same device that sends the mouse press, which itself is
synthesized from the tablet press.

[ChangeLog][QtQuick][PointHandler] If PointHandler is declared with
acceptedButtons: Qt.NoButton it now means that the PointHandler does not
care which buttons are pressed or not pressed: but it ignores
synth-mouse events and responds only to events that come from the
appropriate kind of device. If specific acceptedButtons are set,
synthetic mouse-moves can deactivate it temporarily.

Pick-to: 6.8
Fixes: QTBUG-111336
Change-Id: Ic48565f74d247803e338d2e368a5ef9d249296c2
Reviewed-by: Doris Verria <doris.verria@qt.io>