ListView: Do not create context for section if component is bound
ListView's section needs to inject a "section" property for
compatibility, unless a required property is used for this case.
So far, this made it impossible to use ComponentBehavior: Bound
with sections.
Remedy this by not creating a separate context if we detect that
required properties are used.
Adjust the check in QQuickListViewPrivate::setSectionHelper: If
we don't create a custom context, it might pick up the context
property from somewhere else; but we can then detect that the
used context is internal, and use that check to decide whether
we should call setContextProperyt or setProperty.

Fixes: QTBUG-111433
Pick-to: 6.5
Change-Id: Icd3c138d48f790db92d02c53613539f8f17f5c77
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>