QmlCompiler: Fix logic in generate_As()
We want to communicate the input via the 'in' accumulator and the output
via the 'out' accumulator, not the other way around. It would be a nice
optimization to already read the input as the desired type, but since we
only do QMetaObject::cast() and the "optional" Qvariant conversion for
the "as" operation we cannot guarantee that the lhs can be read that way.

Change-Id: I994e06a02c742ff375cf2edfaaf989adba7d6e90
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>