qt_add_qml_module: Error out if singleton is marked as internal
A singleton cannot be marked as internal – the qmldir parser cannot
even parse such a declaration, and the runtime would have no idea how to
deal with such a type.
The cmake code so far generated two entries – one line where the type is
a singleton, and the other one where the type is declared as internal –
with the end result that the type was not treated as a singleton, at
least not by qmllint.
Once the engine actually supports internal singletons, we would need to
adjust the qmldir file to have singleton and internal in one line.
As that's not supported so far, we now simply error out with a fatal
warning.

Task-number: QTBUG-111532
Change-Id: Idf0843c5548b15f79700a45b4197c256a6dda5cd
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 10a310c089e0d9731067abd6bb56771b51719043)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>