QML: Fix conflicts between QML_ANONYMOUS, QML_UNCREATABLE and namespaces
Even though it makes no sense, we have to allow QML_ANONYMOUS and
QML_UNCREATABLE in the same class since we've documented it. You
cannot trigger the custom message because you'd need a name to attempt
to create an instance of the type. Therefore, the documentation
suggesting this is nonsensical.

In theory you could indeed provide a custom message for namespaces, but
in order to continue supporting this, we would either have to drop the
QmlIsUncreatable mechanism, or we would need a different macro. So,
let's drop this for now. If we really need it, we can introduce a
different macro later.

[ChangeLog][QtQml][Important Behavior Changes] You can no longer
override the "uncreatable" message for QML-exposed namespaces using
QML_UNCREATABLE. This would be incompatible with making the
QML_UNCREATABLE macro safe from unintended inheritance. The latter is
much more important.

Pick-to: 6.5
Fixes: QTBUG-111470
Change-Id: Ib736a2411128d58d939fa2a57d28a0b31c0777e6
Reviewed-by: Robert Griebl <robert.griebl@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>