Models: Make all model roles available for array models
This includes various kinds of objects in lists of various kinds. We
just dynamically create a metaobject that includes all keys we've seen
so far.

For example, you can now pass an array of JavaScript objects as model
to a ComboBox and it will retrieve the correct textRole:

    ComboBox {
        model: [
            { name: "Apple", color: "red" },
            { name: "Orange", color: "orange" },
            { name: "Banana", color: "yellow" }
        ]
        textRole: "color"
    }

This will create the entries "red", "orange", and "yellow".

Amends commit 86794e82ea92cff3b36bbb58f813b625f0230e51.

Change-Id: I3df1f0b43ca1496f7eaed1165731d871c354ac1a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>