Doc: mention the order in which top-level transition animations are run
This is mentioned in the documentation for the transitions property
and in "Playing Animations in Parallel or in Sequence", but is easy to
miss if you only read the detailed description. It's an important
detail that users should know, so add it.

Change-Id: I6f074506ddc8a7b02e6114741d9fadef48cdc8ab
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 0bb5a91d8d9b227dccd93d46b97668c14e20101d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>