manual test: add stylegenerator
Add a manual test that can be used to test the features
of the new qqcstylegenerator tool.

The test will include a style.qtbridge that we can use to
test the tool. Whenever the figma design is updated with
new features/controls, the style.qtbridge file should be
updated as well.

Change-Id: If6dbceaff889ce1e7c2aa2981b02a5d3c7966ba9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>