add a .qt-license-check.exclude file
This file will inform the tst_licenses.pl script
that runs in CI that the some files should not be
included when checking for license headers.

Use this file to inform that the QML template files
that are a part of the qqcstylegenerator tool should
not have license headers. They are meant to be
copied into a users project and become a part
of his own source code (with the freedom to change
them however he wants).

Change-Id: Ifcf85901aef3161ccde5b2cd412f7773bc9edbac
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>