qmlls: implement Find-usages for JS variables
Added the find usages module boilerplate for qmlls and a first
implementation to find usages of JS-declared variables.

A test was added to tst_qmlls_utils to see if the usages can be
found and another one in tst_qmlls_modules to see if the usages are
correctly communicated to the language client (with an extra check
that uris are passed with their "file://"-prefix from qmlls).

Future commits will implement the Find-usages for QmlObject ids,
properties and functions (signals, methods, slots, handlers, ...).

Added QQmlLSUtils::qmlLocationToLspLocation for convertion from
QQmlJS::SourceLocation and QLspSpecification::Range that takes care,
among other things, of 1-based to 0-based line and row conversions.

Changed the VariableDeclarationEntry to contain a IdentifierExpression,
as this allows to find "usages" in the definitions. Adapt the tests to
this change.

Task-number: QTBUG-100084
Task-number: QTBUG-111415
Change-Id: I3a734f3ccf73fa88ed2f28b65ce2582461965ed1
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>