Export QQuickWheelEvent privately
It's mainly intended for QML use cases: both MouseArea and WheelHandler
have wheel(QQuickWheelEvent *) signals. But the QML compiler generates
C++, and other use cases are possible. For the signals to be useful in
C++, their arguments need to be exported. Besides, the other classes in
this file are exported the same way (for example
21b3b74c99e2ee8dedb711ba448066eb308979dc; some were made private in
83e820ba767c2d37600d3536ebcfde18234d2f46).

QQuickWheelEvent was added in f2e1be963f885a6030136591414cdbda26d50695

Pick-to: 6.5
Fixes: QTBUG-111230
Change-Id: I448aa633c7e7deb2e5f12e29bcb1c449be00a63d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>