QQuickTableView: implement SelectedTapped correctly
The current implementation of SelectedTapped allowed
the user to start editing when tapping on the _current_ cell.
This is not correct, SelectedTapped is supposed to (like
the name suggests) let the user start editing when tapping
on a _selected_ cell.

This patch will change the implementation so that
the user can tap on a selected cell to edit it.

Change-Id: Ice5de9250c23f5f57e71077d4f46d1a3f31e9b80
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit bf12f5efa41f4ccd9622b0939a143269e66090cd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>