qv4codegen: Remove static disable_lookup debug aid
This variable, introduced in aabf2196016d96bd171a7d632b6e1cc5622d05ae,
allows to turn of the use of lookups in a few places. However, there is
no apparent need for doing so, and the commit message of the change that
introduced it does not give any reason for it.

As the variable is always false, and given that we run the lookup code
path by default since quite a while, simply remove it.

Change-Id: I62e00739a58fc748a1239087fa0374861ec73e64
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>