Silence some unused-variable and unused-private-field warning
Otherwise, they tend to break the unity build. Since we are
compiling with "-Wattributes", and GCC doesn't warn about
unused variable, having the [[maybe_unused]] leads to an
error, and not having it fails the unity build with other
compilers. As a result, we need the condition to ignore
the flag, and set it again afterward.

Pick-to: 6.5
Task-number: QTBUG-109394
Change-Id: Ia7662a0cfc01de7c9188d37449abdd91e6110fa1
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>