scenegraph: un-blacklist the render autotest on Android
The tst_scenegraph render tests were disabled for Android because they
were failing in CI. This appears to mostly be related to Android devices
having non-integer scaling factors, which causes issues with the render
test as it tests pixel color values derived from QML coordinates.

Most of these issues can be resolved by having the color sample function
round it's floating point values instead of implicitly flooring them
when they are converted to int values.

The render_bug37422 test was failing because the .qsb file was not
handled correctly as test data, so that was converted to use the new
syntax that builds the .qsb file at build time.

The render_OutOfFloatRange test still remains problematic by design on
system with non-integer scaling factors so is disabled when that is the
case.

Fixes: QTBUG-102589
Pick-to: 6.5
Change-Id: I3cae5c9700ef0d5acf2cce46518576de99717c99
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>