Doc: Rename and rephrase pages for QML {script|type} compiler
They describe the user facing concepts, not the actual tools. In the
case of QML script compiler there are two tools that implement it.

Clarify what qmlsc and qmlcachegen actually do and highlight the
differences.

Also, lower case all references to type compiler and script compiler.
They are not product names and therefore should not be upper cased.

Pick-to: 6.5
Change-Id: If800d368c904bd829d88b7376b107b1c5351a1ed
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>