QQuickTableView: only clear selections when selections are enabled
TableView should only modify a selection (which includes clearing
it) when selections are enabled. Otherwise it should leave it to
the application to modify the selection model explicitly.

This can be used to set a selection that "sticks", regardless of
what the user does.

Change-Id: I264946d1b3513846794e1fe12b7d94cfaef050ed
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 9b804c58fcd1085179389ed456afb95957198cf0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>