Support building for Raspberry Pi 1 or Zero
Cross-compiling (using Clang-13) for Raspberry Pi OS on Raspberry Pi 1
and Zero, requires adding "-target arm-linux-gnueabihf" to the compiler
flags, which defines macro __ARM_ARCH_6KZ__ to detect for armv6.

Fixes error:

  qtdeclarative/src/qml/../3rdparty/masm/wtf/Platform.h:312:6: error: "Not supported ARM architecture"

More details: it turns out that the already checked macro
__ARM_ARCH_6ZK__ is a typo with significance because older versions of
GCC only had this. On the other hand __ARM_ARCH_6KZ__ (that this patch
checks for) is the correct spelling, and it's the only one defined by
Clang. Newer versions of GCC define both.

Pick-to: 6.5 6.4 6.2
Change-Id: I60532bfcaa62677f88ed2cff05d872a14c4c2111
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>