Revert value label for the Qt Quick Controls Material slider
The SliderHandle was changed in 0ac8eb67eaa2fca2881caaf439dcac68716f4634
to display the current value of the slider, while the SliderHandle was
being pressed.

This feature should be optional, and we currently don't have a nice API
for hiding it.

Let's look for a way to introduce this feature in Qt 6.6 instead.

Task-number: QTBUG-111355
Change-Id: I3c3f240839d1b69c0fea32d59c8abc8e64e1a184
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 2fcb336698eb812d2e09998e128e35dddcf92fa9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>