QJSEngine: Fix some conversions
The optimization for object-to-string conversion would never trigger
because we were checking the wrong metatype for the PointerToQObject
flag.

Furthermore, we can provide a very simple optimization for the case when
we just want a QObject*. Finally, if we have both types, we can use
is_base_of_v to optimize cases where we are converting to a base class.

Pick-to: 6.5
Task-number: QTBUG-111986
Change-Id: I731fe0398730a2a83222d8c2fdff8aa3d21f7aec
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>