StyleGenerator: change the style generator implementation to use the Figma REST API
This patch will change the implementation of the style generator tool
to use the Figma REST API instead of the qtbridge file. Using the
REST API will let us download and parse the complete style document, and
hence give us all the information we need (which was not the case
with the qtbridge file). After the parsing, we use the REST API to
download the images we need.

To ease the process of adding new controls/Figma templates as
the project makes progress, the description of which atoms we need
to export from Figma has been outsourced to a json config file
rather than being hard-coded from c++.

Task-number: QTBUG-111560
Change-Id: I058a231a61587758d8aec879b1df4454afc166e9
Reviewed-by: Doris Verria <doris.verria@qt.io>