QV4: Mark singletons as deleted before actually deleting them
In cases where a singleton A holds a reference to another singleton B at
the end of the program and if B is destroyed before A, then the update
from B's destruction updates any binding A has on B. If the binding
tries to access B's members then "TypeError: Cannot read property
<property_name> of null" happens.

This fix marks all singletons as deleted before actually deleting them
so that bindings targeting them won't be reevaluated on update.

Randomness, likely due to a random seed in the QHash holding the
singletons changes the order of destruction of the singletons making the
issue sporadic. To account for this in the accompanying test, both
singletons hold a reference to the other.

Fixes: QTBUG-111918
Change-Id: I30450fa08f89f7686c4a41fa46dde2174ce83606
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit c534d6b7b0682d06225e3d801e72f5931b0c504e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>