Update displayText and value of SpinBox (live)
The displayText of a SpinBox is now updated and sanitized (fixup) by
the validator when the user edits the text of the contentItem.

The value of a SpinBox can be updated along the displayText on demand
(set property live to true) when the user edits the text.

[ChangeLog][Controls] The value of a SpinBox can be updated
automatically (live) when the user edits the text. The displayText of a
SpinBox is updated and sanitized automatically.

Fixes: QTBUG-64151
Fixes: QTBUG-85739
Fixes: QTBUG-103205
Change-Id: Id82856190e36d15be2dde8e38091893171f4dad2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>