AnchorChanges: Fix restore logic
The logic to decide whether a state change triggered a size change.
When AnchorChanges did not remove all old anchors, we might end up with
an anchor configuration that changes the size of an item (e.g.,
anchors.left and right being set), even if the new state only adds a
single anchor, e.g. anchors.left.

Moreover, we need to pass the correct dirty flags to setDirty, otherwise
the UI will not actually reflect the state of the Item.
As a driveby, fix an accidental double negation of qt_is_nan for
height.

Fixes: QTBUG-112354
Fixes: QTBUG-106677
Pick-to: 6.5 6.2
Change-Id: I6c456fbdc314672f4fcbe740d941b44183fd5500
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>