Fix stack layout index when loaded asynchronously
Lazy loading of stack layout causes incorrect index value in the
attached property for the child items. This is because,
attachedProperties are created and updated for the child object, after
being added to the stack layout.

This patchset triggers childItemsChanged() to reset child objects of
stack layout with right index value.

Fixes: QTBUG-111902
Pick-to: 6.4 6.5
Change-Id: Iad06d028d0f977189b5da1122904dc754060a609
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>