Enable tst_QQuickCanvasItem on Android
The test mostly passes in the emulator and has generally no reason to
fail in CI. Test functions that should be blacklisted can be added to
BLACKLIST file, or QSKIP'ed if they crash, which none of them do.

Blacklist the test_save function, which fails on Android as we don't
have permission to write to the file system.

Fixes: QTBUG-103257
Task-number: QTBUG-41043
Change-Id: I5ef99a01dc26d9290be91675a5ff15e4de59e812
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 0230ceaec8b0b2ed9ccf7573d2713aa82272cc2b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>